Refactor: 74c0c2e rebase merge
- develop(74c0c2e) -> rebase merge feature/7-new-permission-api
