+++ src/main/java/com/fastcampus/jober/domain/spacewallmember/service/SpaceWallMemberService.java
@@ -5,8 +5,7 @@ import com.fastcampus.jober.domain.member.repository.MemberRepository;
 import com.fastcampus.jober.domain.spacewallmember.domain.SpaceWallMember;
 import com.fastcampus.jober.domain.spacewallmember.dto.SpaceWallMemberRequest;
 import com.fastcampus.jober.domain.spacewallmember.repository.SpaceWallMemberRepository;
-import com.fastcampus.jober.domain.spacewallmember.repository.SpaceWallPermissionRepository;
-import jakarta.persistence.EntityManager;
+import com.fastcampus.jober.domain.spacewallpermission.repository.SpaceWallPermissionRepository;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.stereotype.Service;
