--- src/main/java/com/fastcampus/jober/domain/spacewallmember/service/SpaceWallPermissionService.java
@@ -1,4 +1,0 @@
-package com.fastcampus.jober.domain.spacewallmember.service;
-
-public class SpaceWallPermissionService {
-}
