+++ src/main/java/com/fastcampus/jober/domain/spacewallpermission/domain/SpaceWallPermission.java
@@ -1,11 +1,10 @@
 package com.fastcampus.jober.domain.spacewallpermission.domain;
 
-import com.fastcampus.jober.domain.spacewall.domain.SpaceWall;
-import com.fastcampus.jober.domain.spacewallmember.SpaceWallMember;
+import com.fastcampus.jober.domain.spacewallmember.domain.SpaceWallMember;
 import com.fastcampus.jober.global.constant.Auths;
 import com.fastcampus.jober.global.constant.Type;
 import jakarta.persistence.*;
-import lombok.Data;
+import lombok.*;
 import org.hibernate.annotations.OnDelete;
 import org.hibernate.annotations.OnDeleteAction;
 import org.springframework.data.annotation.CreatedDate;
@@ -14,7 +13,11 @@ import org.springframework.data.annotation.LastModifiedDate;
 import java.time.LocalDateTime;
 
 @Entity
+@Table(name = "space_wall_permission")
 @Data
+@NoArgsConstructor(access = AccessLevel.PROTECTED)
+@AllArgsConstructor
+@Builder
 public class SpaceWallPermission {
 
     @Id
@@ -25,13 +28,7 @@ public class SpaceWallPermission {
     //    @JsonIgnore
     @OneToOne
     @OnDelete(action = OnDeleteAction.CASCADE)
-    @PrimaryKeyJoinColumn
-    private SpaceWall spaceWall;
-
-    //    @JsonIgnore
-    @OneToOne
-    @OnDelete(action = OnDeleteAction.CASCADE)
-    @PrimaryKeyJoinColumn
+    @JoinColumn(name = "space_wall_member_id")
     private SpaceWallMember spaceWallMember;
 
     @Column(nullable = false)
